+++ src/benchmarks/reconstruction/lfhcal_studies/lfhcal_studiesProcessor.cc
@@ -648,7 +648,7 @@ void lfhcal_studiesProcessor::Process(const std::shared_ptr<const JEvent>& event
       m_log->trace("MA cluster {}:\t {} \t {}", iCl, cluster.cluster_E, cluster.cluster_NTowers);
     }
     if (iCl < maxNCluster && enableTreeCluster) {
-      t_lFHCal_clusters_N = (int)iCl;
+      t_lFHCal_clusters_N = iCl;
     }
 
     clusters_calo.clear();
@@ -700,11 +700,10 @@ void lfhcal_studiesProcessor::Process(const std::shared_ptr<const JEvent>& event
       m_log->info("-----> found fEMCClustersF:", fEMCClustersF.size());
       for (const auto cluster : fEMCClustersF) {
         if (iECl < maxNCluster && enableTreeCluster) {
-          t_fEMC_cluster_E[iECl]      = (float)cluster.getEnergy();
+          t_fEMC_cluster_E[iECl]      = cluster.getEnergy();
           t_fEMC_cluster_NCells[iECl] = (int)cluster.getNhits();
-          t_fEMC_cluster_Eta[iECl] =
-              (-1.) * std::log(std::tan((float)cluster.getIntrinsicTheta() / 2.));
-          t_fEMC_cluster_Phi[iECl] = (float)cluster.getIntrinsicPhi();
+          t_fEMC_cluster_Eta[iECl] = (-1.) * std::log(std::tan(cluster.getIntrinsicTheta() / 2.));
+          t_fEMC_cluster_Phi[iECl] = cluster.getIntrinsicPhi();
         }
 
         if (cluster.getEnergy() > highestEEmCl) {
@@ -742,14 +741,14 @@ void lfhcal_studiesProcessor::Process(const std::shared_ptr<const JEvent>& event
                    input_tower_recSav.at(iCell).tower_clusterIDA,
                    input_tower_recSav.at(iCell).tower_clusterIDB);
 
-      t_lFHCal_towers_cellE[iCell]      = (float)input_tower_recSav.at(iCell).energy;
-      t_lFHCal_towers_cellT[iCell]      = (float)input_tower_recSav.at(iCell).time;
+      t_lFHCal_towers_cellE[iCell]      = input_tower_recSav.at(iCell).energy;
+      t_lFHCal_towers_cellT[iCell]      = input_tower_recSav.at(iCell).time;
       t_lFHCal_towers_cellIDx[iCell]    = (short)input_tower_recSav.at(iCell).cellIDx;
       t_lFHCal_towers_cellIDy[iCell]    = (short)input_tower_recSav.at(iCell).cellIDy;
       t_lFHCal_towers_cellIDz[iCell]    = (short)input_tower_recSav.at(iCell).cellIDz;
       t_lFHCal_towers_clusterIDA[iCell] = (short)input_tower_recSav.at(iCell).tower_clusterIDA;
       t_lFHCal_towers_clusterIDB[iCell] = (short)input_tower_recSav.at(iCell).tower_clusterIDB;
-      t_lFHCal_towers_cellTrueID[iCell] = (int)input_tower_recSav.at(iCell).tower_trueID;
+      t_lFHCal_towers_cellTrueID[iCell] = input_tower_recSav.at(iCell).tower_trueID;
     }
 
     event_tree->Fill();
